Optimize test pipeline and workflow sync
