+++ .github/workflows/check-agentic-workflow-sync.yml
@@ -4,6 +4,9 @@ on:
     paths:
       - '.github/agentic-workflows/*.md'
       - '.github/agentic-workflows/*.lock.yml'
+      - '.github/workflows/behavioral-test.md'
+      - '.github/workflows/behavioral-test.lock.yml'
+      - 'scripts/sync-agentic-workflow.js'
 permissions:
   contents: read
 
@@ -12,24 +15,20 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
-      - name: Check behavioral-test .lock.yml exists
+      - name: Sync active workflow files
+        run: node scripts/sync-agentic-workflow.js
+      - name: Install gh-aw CLI
         run: |
-          if [ ! -f .github/agentic-workflows/behavioral-test.lock.yml ]; then
-            echo "❌ behavioral-test.lock.yml not found"
-            echo "Run: gh aw compile .github/agentic-workflows/behavioral-test.md"
-            exit 1
+          if gh extension list | grep -q 'gh-aw'; then
+            gh extension upgrade gh-aw
+          else
+            gh extension install github/gh-aw
           fi
-          echo "✅ behavioral-test.lock.yml exists"
-      - name: Check .md and .lock.yml are in sync
+      - name: Compile active workflow
+        run: gh aw compile --validate --strict .github/workflows/behavioral-test.md
+      - name: Trial active workflow
+        run: gh aw trial .github/workflows/behavioral-test.md --dry-run
+      - name: Verify compiled workflow artifacts exist
         run: |
-          # Extract engine.model from .md frontmatter
-          MD_MODEL=$(grep -A2 'engine:' .github/agentic-workflows/behavioral-test.md | grep 'model:' | awk '{print $2}')
-          # Extract engine.model from .lock.yml
-          LOCK_MODEL=$(grep -A2 'engine:' .github/agentic-workflows/behavioral-test.lock.yml | grep 'model:' | awk '{print $2}')
-
-          if [ "$MD_MODEL" != "$LOCK_MODEL" ]; then
-            echo "❌ Sync mismatch: .md model=$MD_MODEL but .lock.yml model=$LOCK_MODEL"
-            echo "Run: gh aw compile .github/agentic-workflows/behavioral-test.md"
-            exit 1
-          fi
-          echo "✅ .md and .lock.yml are in sync (model: $MD_MODEL)"
+          test -f .github/workflows/behavioral-test.md
+          test -f .github/workflows/behavioral-test.lock.yml
