+++ bunfig.toml
@@ -2,3 +2,7 @@
 # Increase default test timeout from 5s to 15s
 # Some tests invoke bd CLI which requires Dolt server startup
 timeout = 15000
+concurrentTestGlob = "test/scripts/concurrent-*.test.js"
+
+[test.reporter]
+junit = "test-results/test-results.xml"
