+++ lib/commands/test.js
@@ -161,8 +161,8 @@ function getTestCandidatesForChangedFile(file) {
 		return [`test/${relative.replace(/\.js$/, '.test.js')}`];
 	}
 
-	if (file.startsWith('scripts/') && file.endsWith('.js')) {
-		const relative = file.slice('scripts/'.length).replace(/\.js$/, '.test.js');
+	if (file.startsWith('scripts/') && (file.endsWith('.js') || file.endsWith('.sh'))) {
+		const relative = file.slice('scripts/'.length).replace(/\.(js|sh)$/, '.test.js');
 		return [
 			`test/scripts/${relative}`,
 			`test/${relative}`,
@@ -177,9 +177,32 @@ function getTestCandidatesForChangedFile(file) {
 		];
 	}
 
+	if (file.startsWith('.github/agentic-workflows/') || file === '.github/behavioral-test-scores.json') {
+		return [
+			'test/scripts/behavioral-judge.test.js',
+			'test/structural/agentic-workflow-sync.test.js',
+		];
+	}
+
 	if (file.startsWith('.claude/commands/') && file.endsWith('.md')) {
 		return [
 			'test/command-sync-check.test.js',
+			'test/structural/command-sync.test.js',
+		];
+	}
+
+	if (file.startsWith('.forge/')
+		|| file.startsWith('.cursor/')
+		|| file.startsWith('.cline/')
+		|| file.startsWith('.roo/')
+		|| file.startsWith('.kilocode/')
+		|| file.startsWith('.opencode/')
+		|| file.startsWith('.codex/skills/')
+		|| file.startsWith('.github/prompts/')) {
+		return [
+			'test/agent-gaps.test.js',
+			'test/command-sync-check.test.js',
+			'test/scripts/check-agents.test.js',
 			'test/structural/command-sync.test.js',
 		];
 	}
