+++ scripts/test.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 /**
- * Cross-platform test runner for the lefthook pre-push hook.
+ * Cross-platform test runner shared by the pre-push hook and local validation.
  *
- * Runs only the tests affected by the changes being pushed when possible.
- * Falls back to the full suite only for package-level changes.
+ * It runs only the tests affected by known changes when possible and falls back
+ * to the full suite for package-level or unknown-file changes.
  */
 
 const { execFileSync: defaultExecFileSync, spawnSync: defaultSpawnSync } = require('node:child_process');
@@ -22,7 +22,36 @@ const PACKAGE_LEVEL_PATHS = new Set([
   'package-lock.json',
 ]);
 
+const KNOWN_TARGETABLE_PREFIXES = [
+  '.claude/commands/',
+  '.cursor/',
+  '.cline/',
+  '.roo/',
+  '.kilocode/',
+  '.opencode/',
+  '.codex/skills/',
+  '.forge/',
+  '.github/prompts/',
+  '.github/agentic-workflows/',
+  '.github/workflows/',
+  'test/',
+];
+
 const isWindows = process.platform === 'win32';
+
+/**
+ * @typedef {Object} TestExecutionPlan
+ * @property {string[]} changedFiles
+ * @property {boolean} hasUnmappedFiles
+ * @property {boolean} hasUnknownChangedFiles
+ * @property {'targeted'|'full'} mode
+ * @property {string} reason
+ * @property {boolean} runE2E
+ * @property {boolean} runFullSuite
+ * @property {boolean} runTestEnv
+ * @property {boolean} runWorkflowTests
+ * @property {string[]} testTargets
+ */
 
 function detectPackageManager() {
   if (fs.existsSync('bun.lockb') || fs.existsSync('bun.lock')) return 'bun';
@@ -43,13 +72,27 @@ function stripGitHookEnv(sourceEnv = process.env) {
   return env;
 }
 
-function classifyPushTests(projectRoot, execFileSync = defaultExecFileSync) {
-  const changedFiles = getChangedFiles(execFileSync, { sinceUpstream: true });
-  const testTargets = getAffectedTestFiles(projectRoot, execFileSync, fs, { sinceUpstream: true });
+function isKnownTargetablePath(file) {
+  return KNOWN_TARGETABLE_PREFIXES.some((prefix) => file.startsWith(prefix));
+}
+
+function includesWorkflowTarget(testTargets) {
+  return testTargets.some((target) => target === 'test/ci-workflow.test.js'
+    || target === 'test/structural/agentic-workflow-sync.test.js'
+    || target.startsWith('test/workflows/'));
+}
+
+function buildTestExecutionPlan(projectRoot, execFileSync = defaultExecFileSync, options = {}) {
+  const diffOptions = {
+    sinceUpstream: options.sinceUpstream !== false,
+  };
+  const changedFiles = getChangedFiles(execFileSync, diffOptions);
+  const testTargets = getAffectedTestFiles(projectRoot, execFileSync, fs, diffOptions);
 
   let runFullSuite = false;
   let runTestEnv = false;
   let runE2E = false;
+  let runWorkflowTests = includesWorkflowTarget(testTargets);
   let hasUnmappedFiles = false;
   const hasUnknownChangedFiles = changedFiles.length === 0 && testTargets.length === 0;
 
@@ -58,6 +101,7 @@ function classifyPushTests(projectRoot, execFileSync = defaultExecFileSync) {
       runFullSuite = true;
       runTestEnv = true;
       runE2E = true;
+      runWorkflowTests = true;
       break;
     }
 
@@ -71,27 +115,49 @@ function classifyPushTests(projectRoot, execFileSync = defaultExecFileSync) {
       continue;
     }
 
-    if ((file.startsWith('lib/') || file.startsWith('scripts/')) && file.endsWith('.js')) {
+    if ((file.startsWith('lib/') || file.startsWith('scripts/')) && (file.endsWith('.js') || file.endsWith('.sh'))) {
       runTestEnv = true;
+      if (file === 'scripts/behavioral-judge.sh') {
+        runWorkflowTests = true;
+      }
       continue;
     }
 
-    if (file.startsWith('test/') || file.startsWith('.github/workflows/') || file.startsWith('.claude/commands/')) {
+    if (file.startsWith('.github/workflows/') || file.startsWith('.github/agentic-workflows/')) {
+      runWorkflowTests = true;
+    }
+
+    if (isKnownTargetablePath(file)) {
       continue;
     }
 
     hasUnmappedFiles = true;
   }
 
+  const reason = hasUnmappedFiles
+    ? 'unmapped pushed files require full unit coverage'
+    : hasUnknownChangedFiles
+      ? 'changed files could not be resolved safely'
+      : runFullSuite
+        ? 'package-level changes detected'
+        : 'known changes mapped to targeted tests';
+
   return {
     changedFiles,
     hasUnmappedFiles,
     hasUnknownChangedFiles,
+    mode: runFullSuite || hasUnmappedFiles || hasUnknownChangedFiles ? 'full' : 'targeted',
+    reason,
     runE2E,
     runFullSuite: runFullSuite || hasUnmappedFiles || hasUnknownChangedFiles,
     runTestEnv,
+    runWorkflowTests,
     testTargets,
   };
+}
+
+function classifyPushTests(projectRoot, execFileSync = defaultExecFileSync) {
+  return buildTestExecutionPlan(projectRoot, execFileSync, { sinceUpstream: true });
 }
 
 function runCommand(command, args, options = {}, spawnSync = defaultSpawnSync) {
@@ -108,23 +174,17 @@ function runCommand(command, args, options = {}, spawnSync = defaultSpawnSync) {
   return result.status ?? 1;
 }
 
-function runPrePushTests(projectRoot = process.cwd(), deps = {}) {
+function runTestExecutionPlan(plan, deps = {}) {
   const spawnSync = deps.spawnSync || defaultSpawnSync;
-  const execFileSync = deps.execFileSync || defaultExecFileSync;
   const pkgManager = deps.pkgManager || detectPackageManager();
   const env = deps.env || stripGitHookEnv(process.env);
-  const plan = classifyPushTests(projectRoot, execFileSync);
+  const label = deps.label || 'tests';
 
-  console.log(`Running pre-push tests (${pkgManager})...`);
+  console.log(`Running ${label} (${pkgManager})...`);
 
   try {
     if (plan.runFullSuite) {
-      const reason = plan.hasUnmappedFiles
-        ? 'unmapped pushed files require full unit coverage'
-        : plan.hasUnknownChangedFiles
-          ? 'changed files could not be resolved safely'
-          : 'package-level changes detected';
-      console.log(`  Mode: full suite (${reason})`);
+      console.log(`  Mode: full suite (${plan.reason})`);
       const status = runCommand(pkgManager, ['run', 'test'], { env }, spawnSync);
       if (status !== 0) return status;
     } else if (plan.testTargets.length > 0) {
@@ -149,19 +209,37 @@ function runPrePushTests(projectRoot = process.cwd(), deps = {}) {
     return 0;
   } catch (error) {
     console.error('');
-    console.error(`Failed to run tests: ${error.message}`);
+    console.error(`Failed to run ${label}: ${error.message}`);
     console.error('');
     return 1;
   }
 }
 
+function runPrePushTests(projectRoot = process.cwd(), deps = {}) {
+  const execFileSync = deps.execFileSync || defaultExecFileSync;
+  const plan = classifyPushTests(projectRoot, execFileSync);
+  return runTestExecutionPlan(plan, { ...deps, label: 'pre-push tests' });
+}
+
+function runLocalValidationTests(projectRoot = process.cwd(), deps = {}) {
+  const execFileSync = deps.execFileSync || defaultExecFileSync;
+  const plan = buildTestExecutionPlan(projectRoot, execFileSync, { sinceUpstream: true });
+  return runTestExecutionPlan(plan, { ...deps, label: 'local validation tests' });
+}
+
 if (require.main === module) {
-  process.exit(runPrePushTests());
+  const exitCode = process.argv.includes('--validate')
+    ? runLocalValidationTests()
+    : runPrePushTests();
+  process.exit(exitCode);
 }
 
 module.exports = {
+  buildTestExecutionPlan,
   classifyPushTests,
   detectPackageManager,
+  runLocalValidationTests,
   runPrePushTests,
+  runTestExecutionPlan,
   stripGitHookEnv,
 };
