+++ test/ci-workflow.test.js
@@ -1,275 +1,120 @@
 const fs = require('node:fs');
 const path = require('node:path');
-const { describe, test, expect } = require('bun:test');
-const yaml = require('yaml');
+const { describe, test, expect } = require('bun:test');
 
 describe('CI Workflow Configuration', () => {
   const workflowPath = path.join(__dirname, '..', '.github', 'workflows', 'test.yml');
   const workflowContent = fs.readFileSync(workflowPath, 'utf-8');
-  const workflow = yaml.parse(workflowContent);
   const synchronizeSkipCondition = "github.event_name != 'pull_request' || github.event.action != 'synchronize'";
 
-  describe('Coverage Job', () => {
-    test('should have a separate coverage job', () => {
-      expect(workflow.jobs.coverage).toBeTruthy();
-    });
+  function expectSection(sectionName) {
+    expect(workflowContent.includes(`${sectionName}:`)).toBe(true);
+  }
 
-    test('coverage job should run on ubuntu-latest', () => {
-      const coverage = workflow.jobs.coverage;
-      expect(coverage['runs-on']).toBe('ubuntu-latest');
-    });
-
-    test('coverage job should install dependencies', () => {
-      const coverage = workflow.jobs.coverage;
-      const steps = coverage.steps.map(s => s.name);
-      expect(steps.includes('Install dependencies')).toBeTruthy();
-    });
-
-    test('coverage job should run coverage command', () => {
-      const coverage = workflow.jobs.coverage;
-      const coverageStep = coverage.steps.find(s => s.name && s.name.includes('coverage'));
-      expect(coverageStep).toBeTruthy();
-      expect(coverageStep.run && coverageStep.run.includes('bun test --coverage')).toBeTruthy();
-    });
-
-    test('coverage job should upload coverage artifacts', () => {
-      const coverage = workflow.jobs.coverage;
-      const uploadStep = coverage.steps.find(s => s.uses && s.uses.includes('actions/upload-artifact'));
-      expect(uploadStep).toBeTruthy();
-      expect(uploadStep.with && uploadStep.with.name).toBeTruthy();
-      expect(uploadStep.with && uploadStep.with.path).toBeTruthy();
-    });
-
-    test('coverage artifacts should include coverage directory', () => {
-      const coverage = workflow.jobs.coverage;
-      const uploadStep = coverage.steps.find(s => s.uses && s.uses.includes('actions/upload-artifact'));
-      if (uploadStep && uploadStep.with && uploadStep.with.path) {
-        expect(uploadStep.with.path.includes('coverage')).toBeTruthy();
-      }
-    });
-
-    test('coverage job should fail if coverage below threshold', () => {
-      const coverage = workflow.jobs.coverage;
-      const coverageStep = coverage.steps.find(s => s.name && s.name.includes('coverage'));
-      // c8 will fail automatically if below 80% threshold configured in package.json
-      expect(coverageStep).toBeTruthy();
-    });
-  });
-
-  describe('E2E Job', () => {
-    test('should have a separate e2e job', () => {
-      expect(workflow.jobs.e2e).toBeTruthy();
-    });
-
-    test('e2e job should run on ubuntu-latest', () => {
-      const e2e = workflow.jobs.e2e;
-      expect(e2e['runs-on']).toBe('ubuntu-latest');
-    });
-
-    test('e2e job should install dependencies', () => {
-      const e2e = workflow.jobs.e2e;
-      const steps = e2e.steps.map(s => s.name);
-      expect(steps.includes('Install dependencies')).toBeTruthy();
-    });
-
-    test('e2e job should run e2e tests specifically', () => {
-      const e2e = workflow.jobs.e2e;
-      const e2eStep = e2e.steps.find(s => s.name && s.name.includes('E2E'));
-      expect(e2eStep).toBeTruthy();
-      expect(e2eStep.run && e2eStep.run.includes('test/e2e')).toBeTruthy();
-    });
-
-    test('e2e job should setup test fixtures', () => {
-      const e2e = workflow.jobs.e2e;
-      const _fixtureStep = e2e.steps.find(s => s.name && s.name.includes('fixture'));
-      // E2E tests create their own fixtures, but might need setup
-      expect(true).toBeTruthy();
-    });
-  });
-
-  describe('Job Dependencies', () => {
-    test('coverage job should run independently', () => {
-      const coverage = workflow.jobs.coverage;
-      expect(!coverage.needs).toBeTruthy();
-    });
-
-    test('e2e job should run independently', () => {
-      const e2e = workflow.jobs.e2e;
-      expect(!e2e.needs).toBeTruthy();
-    });
-
-    test('all jobs should run in parallel for speed', () => {
-      const test = workflow.jobs.test;
-      const coverage = workflow.jobs.coverage;
-      const e2e = workflow.jobs.e2e;
-
-      expect(!test.needs).toBeTruthy();
-      expect(!coverage.needs).toBeTruthy();
-      expect(!e2e.needs).toBeTruthy();
+  describe('Concurrency', () => {
+    test('workflow cancels older in-progress runs for the same PR or ref', () => {
+      expect(workflowContent.includes('concurrency:')).toBe(true);
+      expect(workflowContent.includes('group: tests-${{ github.event.pull_request.number || github.ref }}')).toBe(true);
+      expect(workflowContent.includes('cancel-in-progress: true')).toBe(true);
     });
   });
 
   describe('Follow-up PR Pushes', () => {
-    test('workflow cancels older in-progress runs for the same PR or ref', () => {
-      expect(workflow.concurrency).toBeTruthy();
-      expect(workflow.concurrency.group).toContain('github.event.pull_request.number || github.ref');
-      expect(workflow.concurrency['cancel-in-progress']).toBe(true);
-    });
-
     test('followup-tests job exists for synchronize events', () => {
-      expect(workflow.jobs['followup-tests']).toBeTruthy();
-      expect(workflow.jobs['followup-tests'].if).toBe("github.event_name == 'pull_request' && github.event.action == 'synchronize'");
+      expectSection('followup-tests');
+      expect(workflowContent.includes("if: github.event_name == 'pull_request' && github.event.action == 'synchronize'")).toBe(true);
     });
 
-    test('followup-tests job runs on a single ubuntu runner', () => {
-      const followup = workflow.jobs['followup-tests'];
-      expect(followup['runs-on']).toBe('ubuntu-latest');
-      expect(!followup.strategy).toBeTruthy();
+    test('followup-tests resolves affected targets through the shared execution planner', () => {
+      expect(workflowContent.includes('name: Resolve affected test targets')).toBe(true);
+      expect(workflowContent.includes('buildTestExecutionPlan')).toBe(true);
+      expect(workflowContent.includes('run_workflow_tests=${plan.runWorkflowTests}')).toBe(true);
+      expect(workflowContent.includes('mode=${plan.mode}')).toBe(true);
     });
 
-    test('followup-tests resolves affected targets before running tests', () => {
-      const followup = workflow.jobs['followup-tests'];
-      const stepNames = followup.steps.map(s => s.name);
-      expect(stepNames.includes('Resolve affected test targets')).toBeTruthy();
-      expect(stepNames.includes('Run targeted unit tests')).toBeTruthy();
-      expect(stepNames.includes('Run single-platform unit suite fallback')).toBeTruthy();
-      expect(stepNames.includes('Run affected e2e tests')).toBeTruthy();
-      expect(stepNames.includes('Run affected edge-case tests')).toBeTruthy();
-    });
-
-    test('followup-tests falls back to full mode for unmapped changed files', () => {
-      const followup = workflow.jobs['followup-tests'];
-      const resolveStep = followup.steps.find((step) => step.name === 'Resolve affected test targets');
-      expect(resolveStep).toBeTruthy();
-      expect(resolveStep.run.includes('unmapped_files=false')).toBeTruthy();
-      expect(resolveStep.run.includes('unmapped_files=true')).toBeTruthy();
-      expect(resolveStep.run.includes('if [[ "$unmapped_files" == "true" ]]')).toBeTruthy();
-    });
-
-    test('followup-tests tracks e2e and test-env changes separately', () => {
-      const followup = workflow.jobs['followup-tests'];
-      const resolveStep = followup.steps.find((step) => step.name === 'Resolve affected test targets');
-      expect(resolveStep).toBeTruthy();
-      expect(resolveStep.run.includes('run_test_env=false')).toBeTruthy();
-      expect(resolveStep.run.includes('run_e2e=false')).toBeTruthy();
-      expect(resolveStep.run.includes('test-env/*')).toBeTruthy();
-      expect(resolveStep.run.includes('test/e2e/*')).toBeTruthy();
-    });
-
-    test('followup-tests runs edge-case checks for library and script changes', () => {
-      const followup = workflow.jobs['followup-tests'];
-      const resolveStep = followup.steps.find((step) => step.name === 'Resolve affected test targets');
-      expect(resolveStep).toBeTruthy();
-      expect(resolveStep.run.includes('lib/*.js|lib/**/*.js')).toBeTruthy();
-      expect(resolveStep.run.includes('scripts/*.js|scripts/**/*.js')).toBeTruthy();
-      expect(resolveStep.run.includes('run_test_env=true')).toBeTruthy();
-    });
-
-    test('full matrix, coverage, e2e, and dashboard skip synchronize events', () => {
-      expect(workflow.jobs.test.if).toBe(synchronizeSkipCondition);
-      expect(workflow.jobs.coverage.if).toBe(synchronizeSkipCondition);
-      expect(workflow.jobs.e2e.if).toBe(synchronizeSkipCondition);
-      expect(workflow.jobs.dashboard.if).toBe(synchronizeSkipCondition);
+    test('followup-tests still runs targeted, fallback, e2e, and edge-case steps', () => {
+      expect(workflowContent.includes('name: Run targeted unit tests')).toBe(true);
+      expect(workflowContent.includes('name: Run single-platform unit suite fallback')).toBe(true);
+      expect(workflowContent.includes('name: Run affected e2e tests')).toBe(true);
+      expect(workflowContent.includes('name: Run affected edge-case tests')).toBe(true);
     });
   });
 
-  describe('Coverage Reporting', () => {
-    test('coverage job should add summary to PR', () => {
-      const coverage = workflow.jobs.coverage;
-      const summaryStep = coverage.steps.find(s => s.name && s.name.includes('summary'));
-      if (summaryStep) {
-        expect(summaryStep.run && summaryStep.run.includes('GITHUB_STEP_SUMMARY')).toBeTruthy();
-      }
+  describe('Fast PR Lane', () => {
+    test('fast PR lane uses four ubuntu shards', () => {
+      expectSection('unit-shard');
+      expect(workflowContent.includes('runs-on: ubuntu-latest')).toBe(true);
+      expect(workflowContent.includes('shard-index: [0, 1, 2, 3]')).toBe(true);
     });
 
-    test('coverage artifacts should be retained for 7 days', () => {
-      const coverage = workflow.jobs.coverage;
-      const uploadStep = coverage.steps.find(s => s.uses && s.uses.includes('actions/upload-artifact'));
-      if (uploadStep && uploadStep.with && uploadStep.with['retention-days']) {
-        expect(uploadStep.with['retention-days'] >= 7).toBeTruthy();
-      }
+    test('fast PR lane keeps platform smoke tests on Node 24 only', () => {
+      expectSection('windows-smoke');
+      expect(workflowContent.includes('runs-on: windows-latest')).toBe(true);
+      expectSection('macos-smoke');
+      expect(workflowContent.includes('runs-on: macos-latest')).toBe(true);
+      expect(workflowContent.includes('node-version: 24')).toBe(true);
+    });
+
+    test('coverage and e2e stay single-platform', () => {
+      expectSection('coverage');
+      expectSection('e2e');
+      expect(workflowContent.includes('name: Code Coverage')).toBe(true);
+      expect(workflowContent.includes('name: E2E Tests')).toBe(true);
+    });
+  });
+
+  describe('Confidence Lane', () => {
+    test('full matrix job is reserved for non-PR runs', () => {
+      expectSection('full-matrix');
+      expect(workflowContent.includes("if: github.event_name != 'pull_request'")).toBe(true);
+      expect(workflowContent.includes('os: [ubuntu-latest, macos-latest, windows-latest]')).toBe(true);
+      expect(workflowContent.includes('node-version: [22, 24]')).toBe(true);
+    });
+
+    test('beads integration is isolated into its own job', () => {
+      expectSection('beads-integration');
+      expect(workflowContent.includes("RUN_BEADS_INTEGRATION: '1'")).toBe(true);
+      expect(workflowContent.includes('name: Run Beads integration tests')).toBe(true);
+      expect(workflowContent.includes('scripts/beads-context.test.js')).toBe(true);
+    });
+  });
+
+  describe('Artifacts and Profiling', () => {
+    test('test jobs upload artifacts and build profiles', () => {
+      expect(workflowContent.includes('scripts/test-profile.js')).toBe(true);
+      expect(workflowContent.includes('uses: actions/upload-artifact@v7')).toBe(true);
+    });
+
+    test('dashboard jobs depend on the appropriate upstream jobs', () => {
+      expectSection('dashboard-pr');
+      expect(workflowContent.includes('needs: [unit-shard, windows-smoke, macos-smoke, coverage, e2e]')).toBe(true);
+      expectSection('dashboard-confidence');
+      expect(workflowContent.includes('needs: [full-matrix, coverage, e2e, beads-integration]')).toBe(true);
+    });
+
+    test('dashboard jobs aggregate artifacts into test-dashboard.json', () => {
+      expect(workflowContent.includes('uses: actions/download-artifact@v5')).toBe(true);
+      expect(workflowContent.includes('scripts/test-dashboard.js')).toBe(true);
+      expect(workflowContent.includes('path: test-dashboard.json')).toBe(true);
     });
   });
 
   describe('Mutation Testing Job', () => {
-    test('mutation job exists in workflow', () => {
-      expect(workflow.jobs.mutation).toBeTruthy();
-    });
-
-    test('mutation job runs on ubuntu-latest', () => {
-      const mutation = workflow.jobs.mutation;
-      expect(mutation['runs-on']).toBe('ubuntu-latest');
-    });
-
-    test('mutation job has workflow_dispatch or schedule condition', () => {
-      const mutation = workflow.jobs.mutation;
-      expect(mutation.if).toBeTruthy();
-      const condition = mutation.if;
-      expect(condition.includes('workflow_dispatch') || condition.includes('schedule')).toBeTruthy();
-    });
-
-    test('mutation job uploads artifacts', () => {
-      const mutation = workflow.jobs.mutation;
-      const uploadStep = mutation.steps.find(s =>
-        s.uses && s.uses.includes('upload-artifact')
-      );
-      expect(uploadStep).toBeTruthy();
+    test('mutation job remains manual or scheduled', () => {
+      expectSection('mutation');
+      expect(workflowContent.includes("if: github.event_name == 'workflow_dispatch' || github.event_name == 'schedule'")).toBe(true);
     });
   });
 
-  describe('Dashboard Job', () => {
-    test('dashboard job exists in workflow', () => {
-      expect(workflow.jobs.dashboard).toBeTruthy();
+  describe('Trigger Layout', () => {
+    test('broad jobs skip synchronize events', () => {
+      const skipConditionOccurrences = workflowContent.split(synchronizeSkipCondition).length - 1;
+      expect(skipConditionOccurrences >= 2).toBe(true);
     });
 
-    test('dashboard job depends on test and coverage', () => {
-      const dashboard = workflow.jobs.dashboard;
-      expect(dashboard.needs).toBeTruthy();
-      const needs = Array.isArray(dashboard.needs) ? dashboard.needs : [dashboard.needs];
-      expect(needs.includes('test')).toBeTruthy();
-      expect(needs.includes('coverage')).toBeTruthy();
-    });
-
-    test('dashboard job uploads artifacts', () => {
-      const dashboard = workflow.jobs.dashboard;
-      const uploadStep = dashboard.steps.find(s =>
-        s.uses && s.uses.includes('upload-artifact')
-      );
-      expect(uploadStep).toBeTruthy();
-    });
-  });
-
-  describe('Schedule Trigger', () => {
-    test('workflow has schedule trigger for mutation testing', () => {
-      expect(workflow.on.schedule).toBeTruthy();
-      expect(Array.isArray(workflow.on.schedule)).toBeTruthy();
-      expect(workflow.on.schedule.length > 0).toBeTruthy();
-      expect(workflow.on.schedule[0].cron).toBeTruthy();
-    });
-  });
-
-  describe('Workflow Structure', () => {
-    test('should have test, coverage, and e2e jobs', () => {
-      expect(workflow.jobs.test).toBeTruthy();
-      expect(workflow.jobs.coverage).toBeTruthy();
-      expect(workflow.jobs.e2e).toBeTruthy();
-    });
-
-    test('test job should still run on multiple platforms', () => {
-      const test = workflow.jobs.test;
-      expect(test.strategy && test.strategy.matrix).toBeTruthy();
-      expect(test.strategy.matrix.os).toBeTruthy();
-      expect(test.strategy.matrix.os.includes('ubuntu-latest')).toBeTruthy();
-      expect(test.strategy.matrix.os.includes('windows-latest')).toBeTruthy();
-      expect(test.strategy.matrix.os.includes('macos-latest')).toBeTruthy();
-    });
-
-    test('coverage and e2e should run on single platform for speed', () => {
-      const coverage = workflow.jobs.coverage;
-      const e2e = workflow.jobs.e2e;
-
-      expect(!coverage.strategy || !coverage.strategy.matrix).toBeTruthy();
-      expect(!e2e.strategy || !e2e.strategy.matrix).toBeTruthy();
+    test('workflow retains schedule and workflow_dispatch triggers', () => {
+      expect(workflowContent.includes('workflow_dispatch:')).toBe(true);
+      expect(workflowContent.includes('schedule:')).toBe(true);
     });
   });
 });
